+++ implementations/cli/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @cotal-ai/cli
 
+## 0.8.3
+
+### Patch Changes
+
+- Updated dependencies [a10ed79]
+  - @cotal-ai/core@0.8.3
+  - @cotal-ai/workspace@0.8.3
+
+## 0.8.2
+
+### Patch Changes
+
+- @cotal-ai/core@0.8.2
+- @cotal-ai/workspace@0.8.2
+
 ## 0.8.1
 
 ### Patch Changes